重大改动啊，d2

+++ Coding_iOS/Controllers/Topic/CSTopicDetailVC.h
@@ -11,7 +11,7 @@
 
 @interface CSTopicDetailVC : UIViewController
 //@property (nonatomic,strong)NSDictionary *topic;
-@property (nonatomic,assign)int topicID;
+@property (nonatomic,assign)NSInteger topicID;
 @end
 
 @interface CSTopTweetDescCell : UITableViewCell
